break decoding out
try embedding in App.tsx
something is wonky
