+++ package.json
@@ -8,10 +8,12 @@
     "@types/jest": "23.3.7",
     "@types/lodash": "^4.14.117",
     "@types/luxon": "^1.4.0",
+    "@types/pako": "^1.0.0",
     "@types/papaparse": "^4.5.4",
     "@types/react": "16.4.18",
     "@types/react-dom": "16.0.9",
-    "d3": "^5.7.0",
+    "@types/d3": "^3.0.0",
+    "d3": "^3.0.0",
     "d3-scale": "^2.1.2",
     "d3-scale-chromatic": "^1.3.3",
     "lodash": "^4.17.4",
