+++ src/planView/QueryPlanGraph.tsx
@@ -1,52 +1,16 @@
 import * as React from "react";
 import * as d3 from "d3";
 import {event as d3Event} from "d3";
+import {
+  buildGraph,
+  QueryNode,
+  QueryPlan,
+} from "./model";
 
 import * as cola from "webcola";
 
 interface QueryPlanGraphProps {
-  data: string;
-}
-
-interface QueryNode extends cola.InputNode {
-  title: string;
-  details: string;
-  type: string;
-  rx: number;
-  ry: number;
-  bounds?: cola.Rectangle;
-  innerBounds?: cola.Rectangle;
-}
-
-interface QueryLink extends cola.Link<QueryNode> {
-  invisible?: boolean;
-  width?: number;
-  route?: {
-    sourceIntersection: cola.Point;
-    targetIntersection: cola.Point;
-    arrowStart: cola.Point;
-  };
-}
-
-interface QueryGroup extends cola.Group {
-  nodeID: string;
-  padding: number;
-}
-
-interface QueryConstraint {
-  type: string;
-  axis: string;
-  offsets: {
-    node: string;
-    offset: number;
-  }[];
-}
-
-interface DataGraph {
-  nodes: QueryNode[];
-  links: QueryLink[];
-  groups: QueryGroup[];
-  constraints: QueryConstraint[];
+  plan: QueryPlan;
 }
 
 // QueryPlanGraph creates an SVG schematic from a JSON representation of a
@@ -58,7 +22,7 @@ export class QueryPlanGraph extends React.Component<QueryPlanGraphProps> {
   zoomRect: d3.Selection<SVGElement>;
 
   onResize = () => {
-    this.redrawPlan(this.props.data);
+    this.redrawPlan(this.props.plan);
   }
 
   constructor(props: QueryPlanGraphProps) {
@@ -75,8 +39,8 @@ export class QueryPlanGraph extends React.Component<QueryPlanGraphProps> {
   }
 
   componentWillReceiveProps(nextProps: QueryPlanGraphProps) {
-    if (nextProps.data !== this.props.data) {
-      this.redrawPlan(nextProps.data);
+    if (nextProps.plan !== this.props.plan) {
+      this.redrawPlan(nextProps.plan);
     }
   }
 
@@ -102,28 +66,8 @@ export class QueryPlanGraph extends React.Component<QueryPlanGraphProps> {
       .call(zoom.event);
   }
 
-  parseJSON(jsonString: string) {
-    const data = JSON.parse(jsonString);
-    ["edges", "nodeNames", "processors"].forEach(attr => {
-      if (!Array.isArray(data[attr])) {
-        throw Error("Invalid query plan provided.");
-      }
-    });
-    return data;
-  }
-
-  redrawPlan(jsonString: string) {
+  redrawPlan(plan: QueryPlan) {
     this.vis.selectAll("*").remove();
-    if (!jsonString) {
-      return;
-    }
-
-    let data;
-    try {
-      data = this.parseJSON(jsonString);
-    } catch (e) {
-      return;
-    }
 
     const containerWidth = this.svg.node().parentElement.clientWidth;
     const containerHeight = this.svg.node().parentElement.clientHeight;
@@ -143,135 +87,7 @@ export class QueryPlanGraph extends React.Component<QueryPlanGraphProps> {
       .flowLayout("y", 80)
       .size([containerWidth, containerHeight]);
 
-    const graph = {} as DataGraph;
-    graph.nodes = [];
-    for (let i = 0; i < data.processors.length; i++) {
-      const p = data.processors[i];
-      p.core.graphNodeIdx = graph.nodes.length;
-      graph.nodes.push({
-        title: p.core.title,
-        details: p.core.details,
-        width: 60,
-        height: 40,
-        rx: 5,
-        ry: 5,
-        type: "core",
-      });
-      for (let j = 0; j < p.inputs.length; j++) {
-        p.inputs[j].graphNodeIdx = graph.nodes.length;
-        graph.nodes.push({
-          title: p.inputs[j].title,
-          details: p.inputs[j].details,
-          width: 60,
-          height: 40,
-          rx: 15,
-          ry: 15,
-          type: "synchronizer",
-        });
-      }
-      for (let j = 0; j < p.outputs.length; j++) {
-        p.outputs[j].graphNodeIdx = graph.nodes.length;
-        graph.nodes.push({
-          title: p.outputs[j].title,
-          details: p.outputs[j].details,
-          width: 60,
-          height: 40,
-          rx: 15,
-          ry: 15,
-          type: "node",
-        });
-      }
-    }
-
-    graph.links = [];
-    for (let i = 0; i < data.edges.length; i++) {
-      let srcNode, destNode;
-
-      const e = data.edges[i];
-      const p1 = data.processors[e.sourceProc];
-      let siblings = 1;
-      if (e.sourceOutput) {
-        srcNode = p1.outputs[e.sourceOutput - 1].graphNodeIdx;
-        if (p1.outputs[e.sourceOutput - 1].title === "by hash") {
-          for (let j = 0; j < data.edges.length; j++) {
-            if (i !== j && data.edges[j].sourceProc === e.sourceProc &&
-                data.edges[j].sourceOutput === e.sourceOutput) {
-              siblings = siblings + 1;
-            }
-          }
-        }
-      } else {
-        srcNode = p1.core.graphNodeIdx;
-      }
-      const p2 = data.processors[e.destProc];
-      if (e.destInput) {
-        destNode = p2.inputs[e.destInput - 1].graphNodeIdx;
-      } else {
-        destNode = p2.core.graphNodeIdx;
-      }
-      let width = 3.0 / siblings;
-      if (width < 1) {
-        width = 1;
-      }
-      graph.links.push({source: srcNode, target: destNode, width: width});
-    }
-
-    // Generate groups.
-    graph.groups = [];
-    for (let i = 0; i < data.nodeNames.length; i++) {
-      graph.groups.push({
-        nodeID: data.nodeNames[i],
-        leaves: [],
-        padding: 15,
-      });
-    }
-    for (let i = 0; i < data.processors.length; i++) {
-      const p = data.processors[i];
-      const n = p.nodeIdx;
-      graph.groups[n].leaves.push(p.core.graphNodeIdx);
-      for (let j = 0; j < p.inputs.length; j++) {
-        graph.groups[n].leaves.push(p.inputs[j].graphNodeIdx);
-      }
-      for (let j = 0; j < p.outputs.length; j++) {
-        graph.groups[n].leaves.push(p.outputs[j].graphNodeIdx);
-      }
-    }
-
-    // Generate constraints to align input synchronizers and output
-    // synchronizers next to the core.
-    graph.constraints = [];
-    for (let i = 0; i < data.processors.length; i++) {
-      const p = data.processors[i];
-      if (p.inputs.length === 0 && p.outputs.length === 0) {
-        continue;
-      }
-      const xConstr = {
-        type: "alignment",
-        axis: "x",
-        offsets: [{node: p.core.graphNodeIdx, offset: 0}],
-      };
-      const yConstr = {
-        type: "alignment",
-        axis: "y",
-        offsets: [{node: p.core.graphNodeIdx, offset: 0}],
-      };
-      const hSpacing = 80, vSpacing = 28 + 10 * graph.nodes[p.core.graphNodeIdx].details.length;
-      for (let j = 0; j < p.inputs.length; j++) {
-        const n = p.inputs[j].graphNodeIdx;
-        xConstr.offsets.push({node: n, offset: hSpacing * (2 * j + 1 - p.inputs.length)});
-        yConstr.offsets.push({node: n, offset: -vSpacing});
-        // These edges are not visible, but they help with the layout.
-        graph.links.push({source: n, target: p.core.graphNodeIdx, invisible: true});
-      }
-      for (let j = 0; j < p.outputs.length; j++) {
-        const n = p.outputs[j].graphNodeIdx;
-        xConstr.offsets.push({node: n, offset: hSpacing * (2 * j + 1 - p.outputs.length)});
-        yConstr.offsets.push({node: n, offset: +vSpacing});
-        // These edges are not visible, but they help with the layout.
-        graph.links.push({source: p.core.graphNodeIdx, target: n, invisible: true});
-      }
-      graph.constraints.push(xConstr, yConstr);
-    }
+    const graph = buildGraph(plan);
 
     const color = d3.scale.category20();
